Question can now be updated.

+++ src/main/java/com/bhtcnpm/website/repository/Exercise/ExerciseQuestionDifficultyRepository.java
@@ -3,5 +3,5 @@ package com.bhtcnpm.website.repository.Exercise;
 import com.bhtcnpm.website.model.entity.ExerciseEntities.ExerciseQuestionDifficulty;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExerciseQuestionDifficultyRepository extends JpaRepository<ExerciseQuestionDifficulty, Long> {
+public interface ExerciseQuestionDifficultyRepository extends JpaRepository<ExerciseQuestionDifficulty, Integer> {
 }
